requirements.txt
requests
google-genai
python-dotenv
tqdm
sentence-transformers
scikit-learn
numpy<2
matplotlib
